add Fish

+++ src/main/java/org/game/MeowPack/FishBucket.java
@@ -5,7 +5,7 @@ import java.awt.*;
 public class FishBucket extends Meow{
     private int fishReleasedSpeed;
 
-    public FishBucket(int meowId, String meowName, float x, float y, int width, int height, int healthPoint, int price, int fishReleasedSpeed) {
+    public FishBucket(int meowId, String meowName, int x, int y, int width, int height, int healthPoint, int price, int fishReleasedSpeed) {
         super(meowId, meowName, x, y, width, height, healthPoint, price);
         this.fishReleasedSpeed = fishReleasedSpeed;
     }
